feat(delete): marginal support for deleting notes
