+++ templates/welcome.html.twig
@@ -44,6 +44,11 @@
             <div id="note-items">
                 {{ _self.noteTitleTemplate() }}
             </div>
+            <ul id="note-menu">
+                <li id="delete-note" class="menu-item" data-action="delete">Delete</li>
+                <li id="copy-note" class="menu-item" data-action="copy">Copy</li>
+            </ul>
+            <div id="note-menu-out-click" class="nicholas"></div>
         </aside>
         <div id="noted">
             <div id="mobile-menu">
